requirements.txt
google-cloud == 0.34.0
google-cloud-storage == 2.8.0
jupyter == 1.0.0
lightning == 2.0.4
matplotlib == 3.7.1
neptune == 1.2.0
numpy == 1.24.2
pandas == 2.0.1
pyarrow==11.0.0
pylint == 2.17.1
PyYAML==6.0
python-dotenv == 1.0.0
scikit-image == 0.20.0
scikit-learn == 1.2.2
torch == 2.0.0
torchmetrics == 1.0.0
torchvision == 0.15.1

fastapi==0.95.1
pydantic==1.10.7
uvicorn==0.22.0